refactor dto 위치 변경 #26
