+++ com.culture-ticket.client.reservation-payment/src/main/java/com/culture_ticket/client/reservation_payment/application/dto/feignclient/PerformanceResponseDto.java
@@ -1,4 +1,4 @@
-package com.culture_ticket.client.reservation_payment.infrastructure.dto;
+package com.culture_ticket.client.reservation_payment.application.dto.feignclient;
 
 import lombok.Getter;
