+++ com.culture-ticket.client.reservation-payment/src/main/java/com/culture_ticket/client/reservation_payment/application/dto/feignclient/TicketRequestDto.java
@@ -1,6 +1,5 @@
-package com.culture_ticket.client.reservation_payment.infrastructure.dto;
+package com.culture_ticket.client.reservation_payment.application.dto.feignclient;
 
-import java.util.List;
 import java.util.UUID;
 import lombok.AccessLevel;
 import lombok.Builder;
