+++ com.culture-ticket.client.reservation-payment/src/main/java/com/culture_ticket/client/reservation_payment/application/service/ReservationService.java
@@ -9,13 +9,13 @@ import com.culture_ticket.client.reservation_payment.common.util.RoleValidator;
 import com.culture_ticket.client.reservation_payment.domain.model.Payment;
 import com.culture_ticket.client.reservation_payment.domain.model.SeatStatus;
 import com.culture_ticket.client.reservation_payment.infrastructure.client.TicketClient;
-import com.culture_ticket.client.reservation_payment.infrastructure.dto.KafkaTicketRequestDto;
-import com.culture_ticket.client.reservation_payment.infrastructure.dto.PerformanceResponseDto;
+import com.culture_ticket.client.reservation_payment.application.dto.kafka.KafkaTicketRequestDto;
+import com.culture_ticket.client.reservation_payment.application.dto.feignclient.PerformanceResponseDto;
 import com.culture_ticket.client.reservation_payment.domain.model.Reservation;
-import com.culture_ticket.client.reservation_payment.infrastructure.dto.SeatResponseDto;
+import com.culture_ticket.client.reservation_payment.application.dto.feignclient.SeatResponseDto;
 import com.culture_ticket.client.reservation_payment.domain.model.SeatPayment;
-import com.culture_ticket.client.reservation_payment.infrastructure.dto.TimeTableResponseDto;
-import com.culture_ticket.client.reservation_payment.infrastructure.dto.UserResponseDto;
+import com.culture_ticket.client.reservation_payment.application.dto.feignclient.TimeTableResponseDto;
+import com.culture_ticket.client.reservation_payment.application.dto.feignclient.UserResponseDto;
 import com.culture_ticket.client.reservation_payment.domain.repository.PaymentRepository;
 import com.culture_ticket.client.reservation_payment.domain.repository.ReservationRepository;
 import com.culture_ticket.client.reservation_payment.domain.repository.SeatPaymentRepository;
