+++ com.culture-ticket.client.reservation-payment/src/main/java/com/culture_ticket/client/reservation_payment/infrastructure/client/PerformanceClient.java
@@ -1,18 +1,16 @@
 package com.culture_ticket.client.reservation_payment.infrastructure.client;
 
 import com.culture_ticket.client.reservation_payment.domain.model.SeatStatus;
-import com.culture_ticket.client.reservation_payment.infrastructure.dto.FeignClientResponseDataDto;
-import com.culture_ticket.client.reservation_payment.infrastructure.dto.FeignClientResponseMessageDto;
-import com.culture_ticket.client.reservation_payment.infrastructure.dto.PerformanceResponseDto;
-import com.culture_ticket.client.reservation_payment.infrastructure.dto.SeatResponseDto;
-import com.culture_ticket.client.reservation_payment.infrastructure.dto.TimeTableResponseDto;
+import com.culture_ticket.client.reservation_payment.application.dto.feignclient.FeignClientResponseDataDto;
+import com.culture_ticket.client.reservation_payment.application.dto.feignclient.FeignClientResponseMessageDto;
+import com.culture_ticket.client.reservation_payment.application.dto.feignclient.PerformanceResponseDto;
+import com.culture_ticket.client.reservation_payment.application.dto.feignclient.SeatResponseDto;
+import com.culture_ticket.client.reservation_payment.application.dto.feignclient.TimeTableResponseDto;
 import java.util.List;
 import java.util.UUID;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestHeader;
