+++ com.culture-ticket.client.reservation-payment/src/main/java/com/culture_ticket/client/reservation_payment/infrastructure/client/UserClient.java
@@ -1,11 +1,10 @@
 package com.culture_ticket.client.reservation_payment.infrastructure.client;
 
-import com.culture_ticket.client.reservation_payment.infrastructure.dto.FeignClientResponseDataDto;
-import com.culture_ticket.client.reservation_payment.infrastructure.dto.UserResponseDto;
+import com.culture_ticket.client.reservation_payment.application.dto.feignclient.FeignClientResponseDataDto;
+import com.culture_ticket.client.reservation_payment.application.dto.feignclient.UserResponseDto;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestHeader;
 
 @FeignClient(name = "user-service")
 public interface UserClient {
